Mercado Pago y Gestion Planes

+++ db/migrate/20210827023252_add_payment_id_to_comercioplanes.rb
@@ -1,5 +1,5 @@
 class AddPaymentIdToComercioplanes < ActiveRecord::Migration[5.2]
   def change
-    add_column :comercioplanes, :payment_id, :integer, default: 0
+    add_column :comercioplanes, :payment_id, :bigint, default: 0
   end
 end
